backend/requirements.txt
fastapi
uvicorn
azure-cognitiveservices-speech
python-dotenv
sounddevice
numpy
wave
sqlalchemy
psycopg2-binary
requests
pyodbc